+++ src/Player.js
@@ -1,0 +1,9 @@
+class Player {
+  constructor(name, symbol) {
+    this.name = name;
+    this.symbol = symbol;
+    this.score = 0;
+  }
+}
+
+export default Player;
